Downloading model scenes and added setup script

+++ GENERATE_SLN_CLEAN.bat
@@ -1,7 +1,10 @@
 @echo off
 
 REM Delete old build folder for a clean build
-rmdir /s /q build
+if exist .\build rmdir /s /q build
 
 REM Generate solution
 call GENERATE_SLN.bat
+
+REM only pause if this is the top-level batch
+if "%~0" == "%~f0" pause
